+++ CHANGELOG.md
@@ -11,17 +11,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   - Extended Data model with New Actions System (NAS) data classes
     (:class:`.bdd.Utilisation`, :class:`.bdd.BaseCiblage` and
-    :class:`.bdd.Ciblage`) and enums (:class:`.bdd.CibleType` and
-    :class:`.bdd.UtilEtat`);
+    :class:`.bdd.Ciblage`) and enums (:class:`.bdd.CibleType`,
+    :class:`.bdd.UtilEtat` and :class:`.bdd.Vote`);
   - Updated ``!fillroles`` to fill BaseCiblages too (new config option
     :attr:`config.max_ciblages_per_action`);
+  - ``!cparti`` now add "vote actions" to all players;
   - New convenience function :func:`.bdd.base.autodoc_DynamicOneToMany`
     for documenting dynamicly loaded one-to-many relationships;
 
 ### Changed
 
   - Updated existing data classes attributes to link to new tables;
+  - Made :attr:`.bdd.Action.base` nullable and added :attr:`.bdd.Action.vote`
+    optionnal attributes to handle votes in a cleaner way; also added
+    :attr:`.bdd.Action.active` attribute to keep track;
+  - New properties :attr:`.bdd.Action.utilisation_ouverte`,
+    :attr:`.bdd.Action.decision`, :attr:`.bdd.Action.is_open` (hybrid)
+    and :attr:`.bdd.Action.is_waiting` (hybrid);
+  - New method :meth:`.bdd.Joueur.action_vote`;
+  - Updated :meth:`.features.open_close.recup_joueurs`, ``!open``,
+    ``!close`` and ``!remind``;
   - New option ``nullable`` for :func:`.bdd.base.autodoc_ManyToOne`;
+  - New class methods for :class:`.bdd.ActionTrigger` enum to get
+    vote-related triggers from :class:`.bdd.Vote` enum ;
   - ``!fillroles`` now post camps descriptions in roles channel;
   - Extended :attr:`.bdd.model_jeu.Camp.description` max length to 1000.